separate types from reducer
