+++ src/store/reducers/userReducer.ts
@@ -1,33 +1,4 @@
-interface UserState {
-  users: any[];
-  loading: boolean;
-  error: null | boolean;
-}
-
-enum UserActionTypes {
-  FETCH_USERS = "FETCH_USERS",
-  FETCH_USERS_SUCCESS = "FETCH_USERS_SUCCESS",
-  FETCH_USERS_ERROR = "FETCH_USERS_ERROR",
-}
-
-interface FetchUsersAction {
-  type: UserActionTypes.FETCH_USERS;
-}
-
-interface FetchUsersSuccessAction {
-  type: UserActionTypes.FETCH_USERS_SUCCESS;
-  payload: any[];
-}
-
-interface FetchUsersErrorAction {
-  type: UserActionTypes.FETCH_USERS_ERROR;
-  payload: string;
-}
-
-type UserAction =
-  | FetchUsersAction
-  | FetchUsersSuccessAction
-  | FetchUsersErrorAction;
+import { UserAction, UserActionTypes, UserState } from "../../types/user";
 
 const initialState: UserState = {
   users: [],
