+++ src/types/user.ts
@@ -1,0 +1,30 @@
+export interface UserState {
+  users: any[];
+  loading: boolean;
+  error: null | boolean;
+}
+
+export enum UserActionTypes {
+  FETCH_USERS = "FETCH_USERS",
+  FETCH_USERS_SUCCESS = "FETCH_USERS_SUCCESS",
+  FETCH_USERS_ERROR = "FETCH_USERS_ERROR",
+}
+
+interface FetchUsersAction {
+  type: UserActionTypes.FETCH_USERS;
+}
+
+interface FetchUsersSuccessAction {
+  type: UserActionTypes.FETCH_USERS_SUCCESS;
+  payload: any[];
+}
+
+interface FetchUsersErrorAction {
+  type: UserActionTypes.FETCH_USERS_ERROR;
+  payload: string;
+}
+
+export type UserAction =
+  | FetchUsersAction
+  | FetchUsersSuccessAction
+  | FetchUsersErrorAction;
